VSTS: stop using DotNetCoreCLI@2 task
The VSTS task tries to be smart and insists on doing its own search
for project files in the tree and running `dotnet` against them
explicitly (at least for the pack command). This is absolutely not
what we want as it completely disregards the build logic in the
traversal project. This is highly annoying.